+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,15 +2,12 @@
 
 using Microsoft.Extensions.Logging;
 
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
@@ -18,11 +15,13 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
     {
         private readonly ILogger<DeleteRestaurantCommandHandler> _logger;
         private readonly IRestaurantsRepository _repository;
+        private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
-        public DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantsRepository repository)
+        public DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantsRepository repository, IRestaurantAuthorizationService restaurantAuthorizationService)
         {
             _logger = logger;
             _repository = repository;
+            _restaurantAuthorizationService = restaurantAuthorizationService;
         }
 
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -31,6 +30,8 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             var restaurant = await _repository.GetByIdAsync(request.Id );
             if (restaurant is null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+            if (!_restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete)) throw new ForbiddenException();
 
             await _repository.DeleteAsync(restaurant);
